+++ internal/udpserver/server_deferred.go
@@ -36,7 +36,7 @@ func (s *Server) deferredConnectAttemptTimeout() time.Duration {
 	return timeout
 }
 
-func (s *Server) processDeferredDNSQuery(ctx context.Context, sessionID uint8, sessionCookie uint8, sequenceNum uint16, downloadCompression uint8, downloadMTUBytes int, assembledQuery []byte) {
+func (s *Server) processDeferredDNSQuery(ctx context.Context, sessionID uint16, sessionCookie uint8, sequenceNum uint16, downloadCompression uint8, downloadMTUBytes int, assembledQuery []byte) {
 	if ctx != nil && ctx.Err() != nil {
 		return
 	}
@@ -78,7 +78,7 @@ func (s *Server) processDeferredDNSQuery(ctx context.Context, sessionID uint8, s
 	}
 }
 
-func (s *Server) finalizeDeferredConnectStream(sessionID uint8, streamID uint16, kind string, outcome string) {
+func (s *Server) finalizeDeferredConnectStream(sessionID uint16, streamID uint16, kind string, outcome string) {
 	if s == nil || sessionID == 0 || streamID == 0 {
 		return
 	}
